# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Enterprise Remote Agent Eval Runner",
    "description": "Enterprise Agent Evaluation Control Plane with Persistent Registry, Version Snapshots, and Zero-SDK Agents.",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/agents": {
      "post": {
        "tags": [
          "Agent Registry"
        ],
        "summary": "Register a new Agent Definition",
        "operationId": "create_agent_api_v1_agents_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Agent Registry"
        ],
        "summary": "Query an Agent by ID (?id=...) or list all agents",
        "operationId": "get_or_list_agents_api_v1_agents_get",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional Agent ID. If omitted, returns all agents.",
              "title": "Id"
            },
            "description": "Optional Agent ID. If omitted, returns all agents."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/AgentResponse"
                    },
                    {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/AgentSummaryResponse"
                      }
                    }
                  ],
                  "title": "Response Get Or List Agents Api V1 Agents Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Agent Registry"
        ],
        "summary": "Delete an Agent by ID (?id=...) with optional force cleanup of local evaluation records",
        "operationId": "delete_agent_api_v1_agents_delete",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Agent ID to delete (required)",
              "title": "Id"
            },
            "description": "Agent ID to delete (required)"
          },
          {
            "name": "force",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Whether to cascade delete local evaluation records.",
              "default": false,
              "title": "Force"
            },
            "description": "Whether to cascade delete local evaluation records."
          },
          {
            "name": "X-Confirm-Name",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Agent exact name to confirm force purge",
              "title": "X-Confirm-Name"
            },
            "description": "Agent exact name to confirm force purge"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentDeleteResponse"
                }
              }
            }
          },
          "400": {
            "description": "Name confirmation mismatch",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DomainErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Agent not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DomainErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Agent has associated launches or active tasks",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DomainErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/experiments/run": {
      "post": {
        "summary": "Run Experiment",
        "operationId": "run_experiment_experiments_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExperimentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperimentResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AgentCreateRequest": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Unique Agent identifier, e.g. banking-agent"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Human-readable name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Detailed description"
          },
          "owner": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner",
            "description": "Team or owner identifier"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "AgentCreateRequest"
      },
      "AgentDeleteResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "ID of deleted Agent"
          },
          "deleted": {
            "type": "boolean",
            "title": "Deleted",
            "description": "Whether the Agent was successfully deleted",
            "default": true
          },
          "launches_deleted": {
            "type": "integer",
            "title": "Launches Deleted",
            "description": "Number of associated experiment launches cleaned up",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Agent deleted successfully"
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "AgentDeleteResponse"
      },
...
      "AgentResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "owner": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "version_count": {
            "type": "integer",
            "title": "Version Count",
            "default": 0
          },
          "latest_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Version"
          },
          "launch_count": {
            "type": "integer",
            "title": "Launch Count",
            "default": 0
          },
          "active_launch_count": {
            "type": "integer",
            "title": "Active Launch Count",
            "default": 0
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "AgentResponse"
      },
      "AgentSummaryResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "owner": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "version_count": {
            "type": "integer",
            "title": "Version Count",
            "default": 0
          },
          "latest_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Version"
          },
          "launch_count": {
            "type": "integer",
            "title": "Launch Count",
            "default": 0
          },
          "active_launch_count": {
            "type": "integer",
            "title": "Active Launch Count",
            "default": 0
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "AgentSummaryResponse"
      },
...
      "DomainErrorResponse": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Machine-readable error code"
          },
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "Human-readable explanation"
          },
          "launch_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Launch Count",
            "description": "Associated launch count if applicable"
          },
          "active_launch_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Launch Count",
            "description": "Active launch count if applicable"
          }
        },
        "type": "object",
        "required": [
          "code",
          "detail"
        ],
        "title": "DomainErrorResponse"
      },
...
      "ExperimentRequest": {
        "properties": {
          "agent_id": {
            "type": "string",
            "title": "Agent Id",
            "default": "banking-agent"
          },
          "agent_version": {
            "type": "string",
            "title": "Agent Version"
          },
          "dataset_name": {
            "type": "string",
            "title": "Dataset Name",
            "default": "banking-agent-regression"
          },
          "experiment_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Experiment Name"
          },
          "max_concurrency": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 50.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Concurrency",
            "description": "Optional concurrency override; if omitted, inherits from AgentVersion"
          }
        },
        "type": "object",
        "required": [
          "agent_version"
        ],
        "title": "ExperimentRequest"
      },
      "ExperimentResult": {
        "properties": {
          "launch_id": {
            "type": "string",
            "title": "Launch Id"
          },
          "agent_id": {
            "type": "string",
            "title": "Agent Id"
          },
          "agent_version": {
            "type": "string",
            "title": "Agent Version"
          },
          "experiment_name": {
            "type": "string",
            "title": "Experiment Name"
          },
          "dataset_run_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dataset Run Url"
          },
          "result": {
            "additionalProperties": true,
            "type": "object",
            "title": "Result"
          }
        },
        "type": "object",
        "required": [
          "launch_id",
          "agent_id",
          "agent_version",
          "experiment_name",
          "result"
        ],
        "title": "ExperimentResult"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}